Fixed warnings with recent Clang (about string formatting and return type mostly)
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileUI@7994 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Headers/ETLayoutItemGroup.h
@@ -65,7 +65,7 @@
 - (ETLayoutItem *) lastItem;
 - (NSInteger) indexOfItem: (id)item;
 - (BOOL) containsItem: (ETLayoutItem *)item;
-- (int) numberOfItems;
+- (NSInteger) numberOfItems;
 - (void) addItems: (NSArray *)items;
 - (void) removeItems: (NSArray *)items;
 - (void) removeAllItems;
